Flatten CI workflow & remove unnecessary CI caching
The CI cache is tied to as single branch, and is not very effective.
It is my estimation that we don't build this often enough that the
caching is worth the extra complexity in the config.

Flatten the workflow so the tests don't have to depend on the check.
This means fewer CI roundtrips to find out that both check and test
jobs fail.